+++ engines/default_engine/assoc.c
@@ -13,7 +13,7 @@
 
 #include "default_engine_internal.h"
 
-#define hashsize(n) ((uint32_t)1<<(n))
+#define hashsize(n) ((size_t)1<<(n))
 #define hashmask(n) (hashsize(n)-1)
 
 ENGINE_ERROR_CODE assoc_init(struct default_engine *engine) {
